feat(invitation): fix ci issues
